Move wireguard_handshake inside wireguard component
