+++ esphome/components/wireguard/wireguard.cpp
@@ -92,6 +92,13 @@ void Wireguard::update() {
   if (this->status_sensor_ != nullptr) {
     this->status_sensor_->publish_state(peer_up);
   }
+
+  if (this->handshake_sensor_ != nullptr) {
+    if (lhs > this->latest_saved_handshake_) {
+      this->latest_saved_handshake_ = lhs;
+      this->handshake_sensor_->publish_state((float) lhs);
+    }
+  }
 }
 
 void Wireguard::dump_config() {
@@ -170,6 +177,7 @@ void Wireguard::set_keepalive(const uint16_t seconds) { this->keepalive_ = secon
 void Wireguard::set_reboot_timeout(const uint32_t seconds) { this->reboot_timeout_ = seconds; }
 void Wireguard::set_srctime(time::RealTimeClock *srctime) { this->srctime_ = srctime; }
 void Wireguard::set_status_sensor(binary_sensor::BinarySensor *sensor) { this->status_sensor_ = sensor; }
+void Wireguard::set_handshake_sensor(sensor::Sensor *sensor) { this->handshake_sensor_ = sensor; }
 
 void Wireguard::disable_auto_proceed() { this->proceed_allowed_ = false; }
